add and delete friends works
