+++ src/components/friends/MyFriendsList.js
@@ -56,7 +56,7 @@ class MyFriends extends Component {
 
                 <section className="favorites-list-container">
                     {this.state.friends.map((friend) => {
-                        return <SearchUserCard {...this.props} key={friend.id} applicant={friend.friend} friend={true} deleteFriend={this.deleteFriend}/>
+                        return <SearchUserCard {...this.props} key={friend.id} applicant={friend.friend} friend={true} deleteFriend={this.deleteFriend} isFriendList={true}/>
                     })}
                 </section>
             </>
