+++ src/components/search/SearchUsers.js
@@ -1,11 +1,34 @@
 // Purpose: To create the USERS search results list component and render result cards
 import React, { Component } from "react"
 import SearchUserCard from "./SearchUserCard";
+import APIManager from "../../modules/APIManager";
+
 class SearchUsers extends Component {
+//     state ={
+//         friends: []
+//     }
 
+//     componentDidMount() {
+//         APIManager.getAllAuth("friends?applicant=True")
+//         .then((friends) => {
+            
+//             this.setState({
+//                 friends: friends
+//             })
+               
+//         })
+//     }
 
-
-
+//   getAllFriends = () => {
+//         APIManager.getAllAuth("friends?applicant=True")
+//         .then((friends) => {
+            
+//             this.setState({
+//                 friends: friends
+//             })
+               
+//         })
+//     }
     render() {
 
         return (
@@ -19,7 +42,7 @@ class SearchUsers extends Component {
                 <section className="company-search-results-container">
                     {this.props.users.map((user) => {
 
-                        return <SearchUserCard {...this.props} key={user.id} applicant={user} />
+                        return <SearchUserCard {...this.props} key={user.id} applicant={user} refreshState={this.getAllFriends} isFriendList={false}/>
 
                     })
                     }
